+++ src/Runtime/__tests__/Runtime-requireModuleOrMock-test.js
@@ -10,7 +10,7 @@
 'use strict';
 
 jest.disableAutomock();
-jest.mock('../../environments/JSDOMEnvironment');
+jest.mock('jest-environment-jsdom');
 
 const path = require('path');
 const normalizeConfig = require('../../config/normalize');
@@ -48,7 +48,7 @@ describe('Runtime', function() {
   beforeEach(function() {
     Runtime = require('../Runtime');
     HasteResolver = require('../../resolvers/HasteResolver');
-    JSDOMEnvironment = require('../../environments/JSDOMEnvironment');
+    JSDOMEnvironment = require('jest-environment-jsdom');
   });
 
   describe('requireModuleOrMock', function() {